refactor: 최종 과제까진 완료
