+++ index_final.html
@@ -10,6 +10,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script type="module" src="/src_final/app/main.tsx"></script>
+<script type="module" src="/src_final/main.tsx"></script>
 </body>
 </html>
